+++ ImageFeed/OAuth2/OAuth2Service.swift
@@ -1,0 +1,2 @@
+
+import Foundation
