+++ ImageFeed/Service/Constants.swift
@@ -1,9 +1,9 @@
 
 import Foundation
 
-let AccessKey = "<ваш Access Key"
-let SecretKey = "<ваш Secret Key"
-let RedirectURI = "<ваш Redirect URI>"
+let AccessKey = "h6OHD3P5qMrGGmFzS6GRBr6xvJcKw4o7kXvstKJuKPk"
+let SecretKey = "aRrzMAK-umgi5Hh0UcKJ-Ne0NWZ-i7sbDCjrikw1Nn0"
+let RedirectURI = "urn:ietf:wg:oauth:2.0:oob"
 let AccessScope = "public+read_user+write_likes"
 let DefaultBaseURL = URL(string: "https://api.unsplash.com")!
 let UnsplashAuthorizeURLString = "https://unsplash.com/oauth/authorize"
